+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - In teacher dropdown, sort by username ([#332](https://gitlab.com/Skn0tt/EntE/issues/332))
 - Allow teachers/managers to unsubscribe from Weekly Summary ([#287](https://gitlab.com/Skn0tt/EntE/issues/287))
+- Translate roles ([#323](https://gitlab.com/Skn0tt/EntE/issues/323))
 - Move Dark Mode out of Beta ([#333](https://gitlab.com/Skn0tt/EntE/issues/333))
 - Hide IDs from user (can still be retrieved from navigation bar) ([#331](https://gitlab.com/Skn0tt/EntE/issues/331))
 - Fixed small typo ([#326](https://gitlab.com/Skn0tt/EntE/issues/326))